rename(gemini): Folder name changed from feature-extract to gemini
